NHOM-9_VU_DUC_NAM/requirements.txt
websockets==11.0.3
pycryptodome==3.19.0
asyncio==3.4.3
flask==2.3.3
flask-cors==3.0.10
flask-socketio==5.3.4